more geometry optimizations to squeeze more DOMs on the GPU
git-svn-id: http://code.icecube.wisc.edu/svn/sandbox/ckopper/clsim@74077 16731396-06f5-0310-8873-f7f720988828

+++ private/clsim/I3CLSimModuleHelper.cxx
@@ -313,7 +313,7 @@ namespace I3CLSimModuleHelper {
                                                              const I3CLSimParticleParameterizationSeries &parameterizationList,
                                                              const std::string &physicsListName,
                                                              double maxBetaChangePerStep,
-                                                             double maxNumPhotonsPerStep,
+                                                             uint32_t maxNumPhotonsPerStep,
                                                              bool multiprocessor)
     {
         I3CLSimParticleToStepConverterGeant4Ptr conv
